Speed up Org mode startup
Hiding drawers and blocks takes the majority of startup time in large
org buffers.  Org mode spends most of this time parsing the blocks and
drawers to be folded in `org-element-at-point'.  Instead of using
generic `org-element-at-point', we can use knowledge about the previous
regexp search involving drawer/blocks regexps.

* lisp/org-cycle.el (org-cycle-hide-drawers): When calling
`org-element-at-point', we are already at the beginning of the drawer
and can directly call `org-element-drawer-parser' from point to the
next heading/EOF.

* lisp/org-fold.el (org-fold-hide-block-all): `org-block-map' searches
blocks with `org-block-regexp'.  The match data can be utilised to
speed-up parsing the block.  We don't need to re-parse from the
beginning of the heading, but simply narrow to the matching region -
we already know where the block is.

* lisp/org.el (org-block-map): Mention that match-data can be used by
the FUNCTION.  This is to avoid breakage of the new
`org-fold-hide-block-all'.