org-element-property: Do not resolve and inline standard properties
* lisp/org-element-ast.el: Add comment that standard properties are
not supposed to be deferred.
(org-element--property):
(org-element-property): Inline queries to standard properties,
yielding `org-element-property-1'.  For other queries, call the
previous body that is not in `org-element--property'.
(org-element-properties): Do not resolve standard properties.
* lisp/org-element.el:
(org-element--cache-find):
(org-element--cache-shift-positions):
(org-element--open-end-p):
(org-element--parse-to):
(org-element-cache-map): Do not call `org-element-property-1'
explicitly.  All the relevant `org-element-property' calls are now
inlined.