Make sure that bult-in and obsolete Org libraries are re-loaded
* lisp/org-load.el (org-require-with-shadowcheck): New macro making
sure that the loaded library is up to date with `load-path'.  If not,
re-load the library version according to current `load-path'.

Update all the top-level requires of internal Org libraries to use the
new macro.

This works around mixed Org versions.

Thanks to Stefan Monnier <monnier@iro.umontreal.ca> for the idea.

Link: https://debbugs.gnu.org/cgi/bugreport.cgi?bug=62762#110