org-element: Parse in-buffer todo keywords within the org-element parser
* lisp/org-element.el (org-element--todo-keyword-setting-re): New
constant matching #+TODO keyword value.
(org-element--compute-todo-keywords): New function and deferred
constant used to parse :TODO, :TYP_TODO, and #+SEQ_TODO, generating
the todo keyword list for the parser.
(org-element-parser-variables): New constant, listing the variables
that affect the parser.  Changing these variables invalidate the
cache.
(org-element-org-data-parser): Use
`org-element--compute-todo-keywords' and store the results in org-data
node.  Store the values of variables listed in
`org-element-parser-variables' within org-data node for reference.
(org-element-cache-version): Bump the cache version.
(org-element--cache-persist-after-read): Do not load persistent cache
when any of the variables affecting the parser is changed in the
current Emacs sesssion.
(org-element-context):
(org-element--headline-parse-title): Use org-data information instead
of in-buffer `org-done-keywords' and `org-todo-regexp'.
(org-element--default-category): Do not try to use `org-category'
variable - this might be dynamically changed in buffer - maintaining
the exact category value in cache is problematic.  Instead, just store
the actual in-buffer value of category according to in-buffer keywords.
* lisp/org.el (org-set-regexps-and-options): Reuse parser data.
(org-get-category): Calculate the actual category according to
in-buffer value from the parser, `org-category' value, or the
fallback.
(org-todo-interpretation): Mark obsolete and move to org-compat.el.