* Add new global attribute: `:export'.
The supported values are: `noexport', `contents' (exports only the
content) or the backends to which it should be exported, separated by
spaces. The name of each backend must be followed by the `+' or `*'
sign, which means that all or only the content must be exported,
respectively. Example: `@foo[:export latex+ odt*]{text}': it will be
exported only to LaTeX (all) and odt (content only).