* lisp/ox.el: Improve export management using an external function.
(org-export-inline-special-block-manage-backends): This new function
controls granular export for all backends. The syntax is modified
somewhat and the `rest' option is added, which is equivalent to the
rest of the backends. Possible values: `noexport', `contents', or backend
name separated by a space (or, simply, `rest'). If the backend name (or
`rest') is followed by the `*' character, only the content is exported.