org-latex-preview: Fix preview visibility with org-cycle
* lisp/org-latex-preview.el (org-latex-preview-auto--org-cycle,
org-latex-preview-auto-mode,
org-latex-preview-auto--handle-post-cursor): When using
`org-cycle', the point is unchanged but the visibility of text
changes.  When there is an open preview overlay at point, this
causes inconsistent behavior when `org-latex-preview-auto-mode' is
enabled.
Fix by checking for invisible text in
`org-latex-preview-auto--handle-post-cursor', and by closing any
open LaTeX preview overlays when folding regions with `org-cycle'.