org-agenda-bulk-action: Do not use global variable name unnecessarily
* lisp/org-agenda-bulk-edit.el (org-agenda-bulk-action): When binding
default time, do not name it `org-overriding-default-time'.  In this
particular case, the DEFAULT-TIME argument passed to `org-read-date'
has the same effect as setting global `org-overriding-default-time':

  (or org-overriding-default-time default-time (org-current-time))

So, naming the argument `org-overriding-default-time' is simply
confusing as we both pass let-bound `org-overriding-default-time'
global and default-time argument at the same time for the same
effect (and unnecessarily confusing for the readers).