org-read-date.el: Document all the global state variables
* lisp/org-read-date.el (org-overriding-default-time): Add docstring.
(org-read-date--default-time):
(org-read-date--default-decoded-time): Expand docstring.
(org-read-date--with-time): Rename from `org-with-time'.
(org-read-date--inactive): Rename from `org-read-date-inactive'.
(org-time-was-given):
(org-read-date-time-was-given): Add alias and docstring.
(org-end-time-was-given):
(org-read-date-end-time-was-given): Add alias and docstring.
(org-read-date): Document setting `org-read-date-time-was-given' and
`org-read-date-end-time-was-given'.  Rename variables as needed.
(org-read-date-analyze): Fix not setting `org-time-was-given' when the
time is 00:XX.  Document setting
`org-read-date-time-was-given' and `org-read-date-end-time-was-given'.
Rename as variables needed.
* lisp/org-timestamp.el (org-timestamp):
* testing/lisp/test-org.el (test-org/org-read-date):
* lisp/org-planning.el (org-add-planning-info): Rename variables.
* lisp/org-obsolete9.8.el (org-with-time):
(org-read-date-inactive): Obsolete variables names.
* testing/lisp/test-org.el (test-org/org-read-date): Remove
let-binding that is no longer necessary.  Fix tests that were relying
upon buggy behavior not recognizing time in 00:40 time input.