org-crypt-check-auto-save: Fix auto-saving multiple Org buffers
* lisp/org-crypt.el (org-crypt--encrypt-entries-all-buffers): New
helper function that encrypts entries as needed in all the Org
buffers.
(org-crypt-check-auto-save): Encrypt entries in _all_ the buffers.
`auto-save-hook' is not evaluated for each buffer, but runs one only.
So, we need to check all Org buffers at once here.
When `org-crypt-disable-auto-save' is set to 'ask, and user answers
"no", still emit "org-decrypt: Decrypting entry with auto-save-mode
enabled..." message.  When `org-crypt-disable-auto-save' is set to
'ask, and user answers "yes", Point user to
`org-crypt-disable-auto-save' that can be customized ('ask is the
default value and users may not know how to disable the prompt).  When
`org-crypt-disable-auto-save' is set to unrecognized value, do not
ignore encryption silently, but throw an error.