org-agenda-get-todos: Use the new agenda-query API
* lisp/org-agenda-search.el (org-agenda-todo-custom-ignore-p): Allow N
argument to be symbol 'future/'past.
(org-agenda--skip-todo-due-to-timestamp): New version of
`org-agenda-check-for-timestamp-as-reason-to-ignore-todo-item' that
allows working with AST nodes.  Use `org-agenda-todo-custom-ignore-p'.
(org-agenda-select-todos): New function returning a list of
headline nodes that should appear in todo agenda blocks.  This is
factored out from `org-agenda-get-todos', stripping the code that
formats headings for agenda display.  It also uses the new API.
(org-agenda-get-todos): Use `org-agenda-select-todos' to form a
list of headings for display.
(org-deadline-close-p): Accept a new optional parameter EPOM.