lisp/org.el: Add proper support for numeric priorities
* doc/org-manual.org: Update the manual to clarify the allowed values for
priorities.

* lisp/org.el (org-priority-highest, org-priority-lowest,
org-priority-default): Update the `defcustom' types so that they perform
validation on the values entered.
(org-priority-numeric-value-regexp): Add a new regexp for valid numeric
priority values for use in input parsing.
(org-priority-regexp): Update the priority cookie regular expression to
fully validate numeric values by using the new
`org-priority-numeric-value-regexp', and update the documentation to match.
(org-priority-valid-cookie-string-p): Add a validation predicate to test
priority cookie strings against the updated `org-priority-regexp' regex.
(org-priority-valid-value-p): Add a validation predicate to test priority
values against allowed values and the currently defined high/low range.
Optionally ignore the defined range in the test so that the predicate can
be used in defcustom validation.
(org-priority-number-p): Add a validation predicate to test if a priority
is in the numeric range (0-64).
(org-priority-to-string): Add a function to provide a consistent string
value from a given priority value.
(org-priority): Update the docstring to clarify acceptable values as well
as the implicit conversion from lower-case to upper-case for alphabetic
values.  Refactor to use the new predicate functions defined in this commit
for validation of values, particularly for double-digit numeric values.
Rename variables used in the function to clarify intent and improve
readability.  Replace bare space characters ("?\ ") with explicit
escapes ("?\s") to improve readability.  Refactor code for interactively
reading new priorities from the user to utilize the predicate functions
defined in this commit and perform additional validation.  Utilize the new
predicate functions to clarify the wrapping logic for `up' and `down'
actions.
(org-entry-put): Replace bare space escape for removal with the `remove'
symbol for consistency.

* testing/lisp/test-org.el: Add unit tests for the new predicates and
format functions related to priority.  Add unit tests for the new
validations for `org-priority-highest', `org-priority-lowest', and
`org-priority-default'.