git-hooks: Add git hook scripts along with instructions for use
* git-hooks/commit-msg: Add commit-msg hook script.  Add a check for two
spaces between sentences.  Update the abort message to reference
`CONTRIBUTE.org' instead of `CONTRIBUTE'.
* git-hooks/commit-msg-files.awk: Add commit-msg-files.awk hook script.
Update the abort message to reference `CONTRIBUTE.org' instead of
`CONTRIBUTE'.
* git-hooks/post-commit: Add post-commit hook script.
* git-hooks/pre-push: Add pre-push hook script.
* git-hooks/prepare-commit-msg: Add prepare-commit-msg hook script.  Update
the abort message to reference `CONTRIBUTE.org' instead of `CONTRIBUTE'.
* mk/targets.mk: Add a `git' and `cleangit' target to install and remove
the new git hook scripts, respectively.  Also add `git' as a dependency for
`all' and `cleangit' as a dependency for `cleanall'.
* CONTRIBUTE.org: Add instructions for utilizing the git hook scripts.