lisp/ox-latex.el: Properly format priority values
Fix LaTeX export so that it properly handles numeric priority values in
addition to characters.

* lisp/ox-latex.el (org-latex-format-headline-default-function): Change
format of priorities to use the `org-priority-to-string' function instead
of assuming a character value.
* testing/lisp/test-ox-latex.el: Fix a typo in the
`test-ox-latex/protect-square-brackets' test so that its name matches the
rest of the tests.  Add unit tests for LaTeX export of headlines with
priorities for both numeric and character values.