org-timestamp-translate: Fix regression causing unconditional <> brackets
* lisp/org.el (org-time-stamp-format): Allow INACTIVE to be
'keep-format to preserve the format as is without stripping/adding
brackets.
(org-timestamp-translate): Use custom timestamp format as is.  It is
indentical to Org 9.5.5 and earlier.  Org 9.6 introduced a regression.
(org-timestamp-custom-formats): Document that brackets are preserved
on export in the docstring.  Clarify that [...] is also stripped.
Mention that variable has no effect on export when
`org-display-custom-times' is nil.
* testing/lisp/test-org.el (test-org/timestamp-translate): Add test
case.

Reported-by: Zeremonienmeister Berber Aab <git@zmberber.com>
Link: https://orgmode.org/list/daffb5e8-c48d-4881-adbe-6dce8043b908@zmberber.com