lisp/oc-bibtex.el: Add default bibliography style variable
* oc-bibtex.el (org-cite-bibtex-export-bibliography): Add
`org-cite-bibtex-bibliography-style' variable to provide a fallback
bibliography style if none is provided.

To render a bibliography bibtex requires a "\bibliographystyle{}" to
be set, omitting a style will create latex documents that fail to
compile.  This variable provides a fallback bibliography style for
when none is provided in the #+CITE_EXPORT options.